Enable FxCop Analyzers in KissTnc directories (#54)
## Description

This is the first of several changes to incrementally enable FxCopAnalyzers in this repository. FxCopAnalyzers allows us to increase code quality of existing code and ensure code quality of new code. Since there are many FxCop violations in this repository, I'm breaking up the large change in to several smaller changes based on project. I will add FxCopAnalyzers to our project Packages.props once the package is enabled everywhere.

This is part of #27 

## Changes

* Adds FxCopAnalyzers to KissTnc and KissTncUnitTests projects
* Fixes FxCop violations in those projects

## Validation

* `dotnet build -c Release` is successful (no violations) on my machine
* Tests are running successfully