Add tests for AprsIsConnection (#96)
## Description

`AprsIsConnection` is missing unit tests. They were prototyped a while back but had not been merged in. This PR updates the branch with the latest in `main` and adds a few tests.

The tests have a timeout while waiting for messages to move. I set the timeout based on multiple runs on my machine, so I suspect they'll be okay, but the timeouts can be upped in the future if they're causing failures.

I also hit a few targets of opportunities for minor tweaks needed in the code.

## Changes

* Add AprsIsConnectionUnitTests
* Tests server login and TCP message received event
* Rename the AprsISLibrary directory to AprsIsConnection to match the project name
* Add a `LoggedIn` property on `AprsIsConnection` to allow the caller to see if login was successful on their connection

## Validation

* Build and tests passing on my machine
* Running APRSsharp console app behaves as expected