Add ReceivedPacket event to AprsIsConnection (#102)
This PR adds a new event that is raised when `AprsIsConnection` receives and successfully decodes a packet. This is different than the existing event that is raised on every single TCP message received from the server. This event fully decodes a packet and passes it to the event handler.

The goal of this change is to offload work from the client by reducing the kind of code the client will have to duplicate. Instead of all clients writing the packet decode call, this can be handled directly by the `AprsIsConnection`, reducing requirements on the client to simply handling the packet that is received.

This resolves #80.