Implement decode on PositionWithoutTimestampWithMessaging (#110)
The last of the remaining scenarios in the `PositionInfo` code is decoding `PacketType.PositionWithoutTimestampWithMessaging`. This PR resolves that issue.

Along with #109, this should allow APRS# to encode and decode any complete weather reports and position packets.

This was a pretty straightforward change as most of the code was already implemented. I added some encode and decode tests (encode of this already worked, only decode needed fixing, but more tests are always good!).

All that was left was to merge the "without timestamp" cases of position decode and set messaging support based on the packet type. This pattern follows the decode of the "with timestamp" types just below in the same file.

This will resolve #92.