Reuse Package Version for APRS-IS Login (#115)
Currently, the AprsSharp version sent to the APRS-IS server is hard-coded in to the connection logic. However, the version used for publishing packages is saved in the package properties (Directory.Build.props sets it across all projects/assemblies).

This changes our code to use reflection to get the assembly version of the AprsIsConnection being used to communicate to the server. Thus eliminating one place to remember to change the version on each release. It should now only be required to update the build properties in the code on each version.