Add AprsIsConnection.State and State Change Events (#116)
This switches from a simple "logged in" Boolean flag to tracking state across the full life of the `AprsIsConnection` object. It also adds an event which can be subscribed to for notifications of when that status changes.

This is especially useful for clients who may wish to know when a connection has been achieved, a login complete, or some error has lead to a disconnect and a new connection is required.