Detect TCP server disconnection (#129)
The current strategy for handling TCP connection interruptions is to rely on an exception on connection disruption. However, if a remote server respectfully closes the connection (as opposed to the connection being interrupted), no exception is generated and the read methods will continue to return empty strings, causing our code to never detect the disconnect and receive empty messages without stopping.

This change pipes through the connection state of the underlying `TcpClient` to check connection on each cycle of the receive loop. This allows the code to detect a disconnection by the server and set the state of the `AprsIsConnection` to disconnected to match.