Rename AprsIsConnection to AprsIsClient (#126)
This changes `AprsIsConnection` project and class to `AprsIsClient` to match more closely with .NET naming conventions (`TcpClient`, `HttpClient`, etc.). It's also a slightly more descriptive name as this is coded exclusively as a client implementation, not a server, so it's arguably not a generic connection between APRS-IS nodes.

This also brings the code in line with the name of the nuget package [AprsSharp.AprsIsClient](https://www.nuget.org/packages/AprsSharp.AprsIsClient/).