Resolve double "filter" in default filtering (#168)
## Description

As reported in #162 by @ShawnStoddard (thanks!), the default filter
string in the CLI includes the word `filter`, which is later appended by
the AprsIsClient. This PR removes that additional/unnecessary string and
updates the docs to ensure it's clear that the filter string is just the
logic string, not the extra "filter" word at the front.

This resolves #162

## Changes

* Remove additional "filter" word in default filter value
* Update docstrings for clarity

## Validation

* Manual testing ensures APRS-IS filtering is working correctly
* Manual testing ensures new output appears in `--help` flag to CLI