Remove Required TcpConnection Object from TcpTnc (#181)
## Description

The `TcpConnection` object exists to allow easy testing. However, it
isn't super useful for users of the package..and requires extra code and
understanding extra concepts. This PR adds the ability of `TcpTnc` to
manage its own `TcpConnection` while preserving the ability to pass in
an object for testing.

## Changes

* Allow `TcpTnc` construction without a `TcpConnection` object (managed
inside the `TcpTnc` object directly)
* Update in CLI app to use the simpler constructor

## Validation

* CI passes
* Manual test against a TCP KISS TNC is functioning well